+++ src/main/java/org/ccctc/colleaguedmiclient/service/DmiDataService.java
@@ -8,8 +8,10 @@ import org.apache.commons.logging.LogFactory;
 import org.ccctc.colleaguedmiclient.exception.DmiServiceException;
 import org.ccctc.colleaguedmiclient.model.CddEntry;
 import org.ccctc.colleaguedmiclient.model.ColleagueData;
+import org.ccctc.colleaguedmiclient.model.ElfTranslateTable;
 import org.ccctc.colleaguedmiclient.model.EntityMetadata;
 import org.ccctc.colleaguedmiclient.model.SessionCredentials;
+import org.ccctc.colleaguedmiclient.model.Valcode;
 import org.ccctc.colleaguedmiclient.transaction.data.SelectRequest;
 import org.ccctc.colleaguedmiclient.transaction.data.SelectResponse;
 import org.ccctc.colleaguedmiclient.transaction.data.SingleKeyRequest;
@@ -41,6 +43,12 @@ public class DmiDataService {
 
     private final Log log = LogFactory.getLog(DmiDataService.class);
     private final DmiService dmiService;
+
+    private final List<String> valcodeColumns = Arrays.asList("VAL.INTERNAL.CODE", "VAL.EXTERNAL.REPRESENTATION",
+            "VAL.ACTION.CODE.1", "VAL.ACTION.CODE.2");
+
+    private final List<String> elfTranslateColumns = Arrays.asList("ELFT.DESC", "ELFT.COMMENTS", "ELFT.ORIG.CODE.FIELD",
+            "ELFT.ORIG.CODES", "ELFT.NEW.CODES", "ELFT.ACTION.CODES.1", "ELFT.ACTION.CODES.2");
 
     /**
      * Entity metadata service used to translate field names and data types from a response
@@ -303,14 +311,89 @@ public class DmiDataService {
 
         logSend("selectKeys", viewName, null, null, criteria);
 
-        DmiTransaction dmiReponse = dmiService.send(request);
-        SelectResponse selectResponse = SelectResponse.fromDmiTransaction(dmiReponse);
+        DmiTransaction dmiResponse = dmiService.send(request);
+        SelectResponse selectResponse = SelectResponse.fromDmiTransaction(dmiResponse);
 
         logReceive("selectKeys", viewName, selectResponse.getKeys().length);
 
         return selectResponse.getKeys();
     }
 
+
+    /**
+     * Get a valcode record
+     *
+     * @param appl Application
+     * @param key  Valcode key
+     * @return Valcode
+     */
+    public Valcode valcode(String appl, String key) {
+        ColleagueData data = singleKey(appl, appl + ".VALCODES", valcodeColumns, key);
+        if (data != null) return processValcode(data);
+        return null;
+    }
+
+
+    /**
+     * Get multiple valcode records
+     *
+     * @param appl Application
+     * @param keys Valcode keys
+     * @return Valcodes
+     */
+    public List<Valcode> valcodes(String appl, Iterable<String> keys) {
+        List<ColleagueData> data = batchKeys(appl, appl + ".VALCODES", valcodeColumns, keys);
+
+        if (data != null) {
+            List<Valcode> result = new ArrayList<>();
+
+            for (ColleagueData d : data) {
+                Valcode v = processValcode(d);
+                if (v != null) result.add(v);
+            }
+
+            return result;
+        }
+
+        return null;
+    }
+
+
+    /**
+     * Get an ELF Translation Table record
+     *
+     * @param key  ELF Translation Table key
+     * @return ELF Translation Table
+     */
+    public ElfTranslateTable elfTranslationTable(String key) {
+        ColleagueData data = singleKey("CORE", "ELF.TRANSLATE.TABLES", elfTranslateColumns, key);
+        if (data != null) return processElfTranslationTable(data);
+        return null;
+    }
+
+
+    /**
+     * Get multiple ELF Translation Table records
+     *
+     * @param keys ELF Translation Table keys
+     * @return ELF Translation Tables
+     */
+    public List<ElfTranslateTable> elfTranslationTables(Iterable<String> keys) {
+        List<ColleagueData> data = batchKeys("CORE", "ELF.TRANSLATE.TABLES", elfTranslateColumns, keys);
+
+        if (data != null) {
+            List<ElfTranslateTable> result = new ArrayList<>();
+
+            for (ColleagueData d : data) {
+                ElfTranslateTable v = processElfTranslationTable(d);
+                if (v != null) result.add(v);
+            }
+
+            return result;
+        }
+
+        return null;
+    }
 
     /**
      * Process a DMI response for this data request. The result is a list of records with the field names
@@ -327,7 +410,11 @@ public class DmiDataService {
 
         List<ColleagueData> result = new ArrayList<>();
 
-        EntityMetadata entityMetadata = entityMetadataService.get(appl, viewName);
+        String entityName = viewName;
+        if (viewName.length() >= appl.length() + 1 && (appl + ".").equals(viewName.substring(0, appl.length() + 1)))
+            entityName = "appl." + viewName.substring(appl.length() + 1);
+
+        EntityMetadata entityMetadata = entityMetadataService.get(appl, entityName);
         DataResponse dataResponse = DataResponse.fromDmiTransaction(dmiResponse);
 
         if (dataResponse.getOrder().size() > 0) {
@@ -379,6 +466,76 @@ public class DmiDataService {
 
         if (stringValue != null)
             return CddUtils.convertToValue(stringValue, cddEntry);
+
+        return null;
+    }
+
+
+    /**
+     * Process Valcode data
+     *
+     * @param data Data
+     * @return Valcode
+     */
+    private Valcode processValcode(ColleagueData data) {
+        if (data != null) {
+            List<Valcode.Entry> entries = new ArrayList<>();
+
+            String[] internalCodes = (String[]) data.getValues().get("VAL.INTERNAL.CODE");
+            String[] externalRepresentations = (String[]) data.getValues().get("VAL.EXTERNAL.REPRESENTATION");
+            String[] action1s = (String[]) data.getValues().get("VAL.ACTION.CODE.1");
+            String[] action2s = (String[]) data.getValues().get("VAL.ACTION.CODE.2");
+
+            int count = (internalCodes != null) ? internalCodes.length : 0;
+            for (int x = 0; x < count; x++) {
+                String i = internalCodes[x];
+                String e = (externalRepresentations != null && externalRepresentations.length > x) ? externalRepresentations[x] : null;
+                String a1 = (action1s != null && action1s.length > x) ? action1s[x] : null;
+                String a2 = (action2s != null && action2s.length > x) ? action2s[x] : null;
+                Valcode.Entry entry = new Valcode.Entry(i, e, a1, a2);
+                entries.add(entry);
+            }
+
+            return new Valcode(data.getKey(), entries);
+        }
+
+        return null;
+    }
+
+
+    /**
+     * Process ELF Translation Table data
+     *
+     * @param data Data
+     * @return ELF Translation Table
+     */
+    private ElfTranslateTable processElfTranslationTable(ColleagueData data) {
+        if (data != null) {
+            List<ElfTranslateTable.Entry> translations = new ArrayList<>();
+
+            String description = (String) data.getValues().get("ELFT.DESC");
+            String origCodeField = (String) data.getValues().get("ELFT.ORIG.CODE.FIELD");
+            String newCodeField = (String) data.getValues().get("ELFT.NEW.CODE.FIELD");
+            String[] comments = (String[]) data.getValues().get("ELFT.COMMENTS");
+            String[] originalCodes = (String[]) data.getValues().get("ELFT.ORIG.CODES");
+            String[] newCodes = (String[]) data.getValues().get("ELFT.NEW.CODES");
+            String[] action1s = (String[]) data.getValues().get("ELFT.ACTION.CODES.1");
+            String[] action2s = (String[]) data.getValues().get("ELFT.ACTION.CODES.2");
+
+            int count = (originalCodes != null) ? originalCodes.length : 0;
+            for (int x = 0; x < count; x++) {
+                String o = originalCodes[x];
+                String n = (newCodes != null && newCodes.length > x) ? newCodes[x] : null;
+                String a1 = (action1s != null && action1s.length > x) ? action1s[x] : null;
+                String a2 = (action2s != null && action2s.length > x) ? action2s[x] : null;
+                ElfTranslateTable.Entry entry = new ElfTranslateTable.Entry(o, n, a1, a2);
+                translations.add(entry);
+            }
+
+            List<String> c = (comments != null) ? Arrays.asList(comments) : null;
+
+            return new ElfTranslateTable(data.getKey(), description, c, origCodeField, newCodeField, translations);
+        }
 
         return null;
     }