+++ release-notes.md
@@ -1,5 +1,10 @@
 # Release Notes #
 
+__1.0.6__
+
+* Fix for data reads from application based files (ie ST.VALCODES)
+* New methods in DmiDataService for reading valcodes and ELF translation tables
+
 __1.0.5__
 
 * Fix for blank lines at end of DMI transaction on certain error conditions